# swagger.json
{
"openapi": "3.0.0",
"paths": {
...
"/api/users/{id}": {
"get": {
"operationId": "UsersController_findById",
"parameters": [
{
"name": "id",
"required": true,
"in": "path",
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Data user"
},
"404": {
"description": "User tidak ditemukan"
}
},
"summary": "Get user by ID",
"tags": [
"Users"
]
}
},
"/api/users/profile": {
"put": {
"operationId": "UsersController_updateProfile",
"parameters": [],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/UpdateProfileDto"
}
}
}
},
"responses": {
"200": {
"description": "Profile berhasil diupdate"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Update profile user",
"tags": [
"Users"
]
},
"delete": {
"operationId": "UsersController_deleteAccount",
"parameters": [],
"responses": {
"200": {
"description": "Akun berhasil dihapus"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Hapus akun user",
"tags": [
"Users"
]
}
},
"/api/users/xp": {
"put": {
"operationId": "UsersController_updateXP",
"parameters": [],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"xp": {
"type": "number",
"example": 50
}
}
}
}
}
},
"responses": {
"200": {
"description": "XP berhasil diupdate"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Update XP user",
"tags": [
"Users"
]
}
},
...
"/api/progress": {
"get": {
"operationId": "ProgressController_getUserProgress",
"parameters": [],
"responses": {
"200": {
"description": "Data progress user"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Get semua progress user",
"tags": [
"Progress"
]
}
},
"/api/progress/stats": {
"get": {
"operationId": "ProgressController_getStats",
"parameters": [],
"responses": {
"200": {
"description": "Data statistik user"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Get statistik user (total lesson, XP, dll)",
"tags": [
"Progress"
]
}
},
"/api/progress/lesson/{lessonId}": {
"get": {
"operationId": "ProgressController_getLessonProgress",
"parameters": [
{
"name": "lessonId",
"required": true,
"in": "path",
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Data progress lesson"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Get progress lesson tertentu",
"tags": [
"Progress"
]
}
},
"/api/progress/start": {
"post": {
"operationId": "ProgressController_startLesson",
"parameters": [],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"levelId": {
"type": "string",
"example": "507f1f77bcf86cd799439011"
},
"lessonId": {
"type": "string",
"example": "507f1f77bcf86cd799439012"
}
}
}
}
}
},
"responses": {
"201": {
"description": "Lesson dimulai"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Mulai lesson baru",
"tags": [
"Progress"
]
}
},
"/api/progress/complete": {
"post": {
"operationId": "ProgressController_completeLesson",
"parameters": [],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"lessonId": {
"type": "string",
"example": "507f1f77bcf86cd799439012"
},
"score": {
"type": "number",
"example": 85
},
"timeSpent": {
"type": "number",
"example": 120,
"description": "Dalam detik"
},
"detailedProgress": {
"type": "object",
"example": {
"correctAnswers": 8,
"totalQuestions": 10
}
}
}
}
}
}
},
"responses": {
"200": {
"description": "Lesson selesai"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Selesaikan lesson",
"tags": [
"Progress"
]
}
},
"/api/quiz/history": {
"get": {
"operationId": "QuizController_getHistory",
"parameters": [],
"responses": {
"200": {
"description": "Quiz history"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Get riwayat quiz user",
"tags": [
"Quiz"
]
}
},
"/api/quiz/{id}/results": {
"get": {
"operationId": "QuizController_getResult",
"parameters": [
{
"name": "id",
"required": true,
"in": "path",
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Data hasil quiz"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Get hasil quiz by ID",
"tags": [
"Quiz"
]
}
},
"/api/quiz/submit": {
"post": {
"operationId": "QuizController_submit",
"parameters": [],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/SubmitDto"
}
}
}
},
"responses": {
"201": {
"description": "Quiz berhasil disubmit"
}
},
"security": [
{
"JWT-auth": []
}
],
"summary": "Submit jawaban quiz",
"tags": [
"Quiz"
]
}
},
...
},
"info": {
"title": "SignMaster API",
"description": "API untuk aplikasi belajar bahasa isyarat SignMaster",
"version": "1.0",
"contact": {}
},
...
"servers": [],
"components": {
"securitySchemes": {
"JWT-auth": {
"scheme": "bearer",
"bearerFormat": "JWT",
"type": "http",
"name": "JWT",
"description": "Enter JWT token",
"in": "header"
}
},
"schemas": {
"UpdateProfileDto": {
"type": "object",
"properties": {
"displayName": {
"type": "string",
"example": "John Doe"
},
"bio": {
"type": "string",
"example": "Saya sedang belajar bahasa isyarat"
}
}
},
...
"AnswerDto": {
"type": "object",
"properties": {
"questionId": {
"type": "string",
"example": "507f1f77bcf86cd799439011"
},
"userAnswer": {
"type": "string",
"example": "A"
},
"timeSpent": {
"type": "number",
"example": 5
}
},
"required": [
"questionId",
"userAnswer",
"timeSpent"
]
},
"MistakeDto": {
"type": "object",
"properties": {
"position": {
"type": "number",
"example": 2
},
"expectedLetter": {
"type": "string",
"example": "B"
},
"detectedLetter": {
"type": "string",
"example": "D"
}
},
"required": [
"position",
"expectedLetter",
"detectedLetter"
]
},
"CameraChallengeDto": {
"type": "object",
"properties": {
"word": {
"type": "string",
"example": "HALO"
},
"mistakes": {
"type": "array",
"items": {
"$ref": "#/components/schemas/MistakeDto"
}
},
"averageConfidence": {
"type": "number",
"example": 0.85
}
},
"required": [
"word",
"mistakes",
"averageConfidence"
]
},
"SubmitDto": {
"type": "object",
"properties": {
"lessonId": {
"type": "string",
"example": "507f1f77bcf86cd799439011"
},
"quizType": {
"type": "string",
"example": "quiz",
"enum": [
"practice",
"quiz",
"camera-challenge"
]
},
"answers": {
"type": "array",
"items": {
"$ref": "#/components/schemas/AnswerDto"
}
},
"cameraChallenge": {
"$ref": "#/components/schemas/CameraChallengeDto"
}
},
"required": [
"lessonId",
"quizType",
"answers"
]
}
}
}
}